Changes: # Debugged both Evalutations files

+++ sandbox/EvaluateClassifierUncertainty.py
@@ -2,7 +2,7 @@ import torch
 from torch import nn
 import torch.nn.modules
 import torch.optim as optim
-import torch.functional as F
+import torch.nn.functional as F
 import torchvision
 import torchvision.transforms
 
@@ -82,7 +82,7 @@ class EvaluateClassifierUncertainty:
         for _ in range(num_iter):
             Zs_params = torch.ones(samples_per_iter, dimZ*2)
             reconstructed_Xs = task_model.sample_and_decode(Zs_params)
-            true_Ys = torch.ones(samples_per_iter) * task_id # these are the labels for the generated pictures
+            true_Ys = torch.ones(samples_per_iter, dtype=torch.long) * task_id # these are the labels for the generated pictures
             cross_entropies = F.cross_entropy(self.classifier(reconstructed_Xs), true_Ys, reduction='none')
             loss_mu += torch.mean(cross_entropies) / num_iter
             loss_var += torch.mean((cross_entropies - loss_mu)**2)
